add api and projections for read model

+++ EventSourcingDemo/Order.cs
@@ -1,3 +1,3 @@
 ﻿namespace EventSourcingDemo;
 
-public record Order(string ProductName, int ProductId, int Quantity, int Price, string Comment);
+public record Order(string ProductName, int ProductId, int Quantity, double Price, string Comment);
